feat: add get( ) handling GET method
